test nro "no sé"

+++ src/UberFrba/Registro Viajes/RegistroViajeForm.cs
@@ -101,7 +101,7 @@ namespace UberFrba.Registro_Viajes
             else
                 return null;
 
-            nuevo.km_viaje = (int)kmNumericUpDown.Value;
+            nuevo.km_viaje = (double)kmNumericUpDown.Value;
             nuevo.inicio_date = beginDateTimePicker.Value;
             nuevo.fin_date = endDateTimePicker.Value;
